% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{example.project}
\alias{example.project}
\title{Create Example project}
\usage{
example.project()
}
\description{
Create Example project
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
