% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.brks}
\alias{fx.brks}
\title{get breaks from a raster based on a given number of breaks. default is 32}
\usage{
fx.brks(r, n = 32)
}
\arguments{
\item{r}{raster}

\item{n}{Number of breaks (default is 32)}
}
\value{
brks The breaks
}
\description{
get breaks from a raster based on a given number of breaks. default is 32
}

