% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.CheckInstallLoadPkgs}
\alias{fx.CheckInstallLoadPkgs}
\title{Function to check and install pacakges}
\usage{
fx.CheckInstallLoadPkgs(x)
}
\arguments{
\item{x}{vector of packages to check and install}
}
\description{
Function to check and install pacakges
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
