% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.click.profile}
\alias{fx.click.profile}
\title{Draw an elevation profile line on a raster}
\usage{
fx.click.profile(r)
}
\arguments{
\item{r}{raster}
}
\value{
table of elevations
}
\description{
Draw an elevation profile line on a raster
}
