% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.click}
\alias{fx.click}
\title{Click and plot clicked points while still clicking}
\usage{
fx.click(r, n = 100, xy = TRUE, PolyLine = FALSE)
}
\arguments{
\item{r}{raster}

\item{n}{number of clicks}
}
\value{
cp
}
\description{
Click and plot clicked points while still clicking
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
