% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.edit.terrain}
\alias{fx.edit.terrain}
\title{Function to edit a DEM(depends on other functions above)}
\usage{
fx.edit.terrain(dem = "dem.tif", nlines = 4, zmanual = FALSE,
  OutPrefix = "Result",
  coord.sys = "+proj=utm +zone=32 +ellps=GRS80 +units=m +no_defs",
  plot.shapes = FALSE, male.pts = NULL, other.shps = NULL, xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{dem}{Path to a raster withe levation data (tif).}

\item{nlines}{NUmber of editing lines to be drawn for cutting/filling terrain}

\item{zmanual}{TRUE/FALSE determines whether elevations fro start end endpoints to the editing lines are enteered manually or read from the terrain}

\item{OutPrefix}{a prefix for the output filenames}
}
\value{
A list with elements: (1) Drawn interpolation lines and (2) bounding polygon polys.df and (3) Edited terrain model dem
}
\description{
Function to edit a DEM(depends on other functions above)
}
\examples{
fx.edit.terrain("dem.tif",nlines=4,zmanual=F,OutPrefix="Result")
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

