% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.ExtractByMask}
\alias{fx.ExtractByMask}
\title{Extract by mask}
\usage{
fx.ExtractByMask(r, clicks = 100)
}
\arguments{
\item{r}{raster}

\item{clicks}{number of clicks to make a mask polygon (maks 100)}
}
\value{
out
}
\description{
Extract by mask
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
