% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.PolygonFromRasterClicks}
\alias{fx.PolygonFromRasterClicks}
\title{Draw Polygon by cklicking in a plotted Raster}
\usage{
fx.PolygonFromRasterClicks(r, clicks = 100)

fx.PolygonFromRasterClicks(r, clicks = 100)
}
\arguments{
\item{r}{raster}

\item{clicks}{number of clicks}

\item{r}{raster}

\item{clicks}{number of clicks}
}
\value{
shp

shp
}
\description{
Draw Polygon by cklicking in a plotted Raster

Draw Polygon by cklicking in a plotted Raster
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}

Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
