% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.Terrain2Curves}
\alias{fx.Terrain2Curves}
\title{Plots volume-elevation curve and hypsographic curve}
\usage{
fx.Terrain2Curves(zlim = NULL, clip2shp = TRUE, zones = 10,
  change.dir = TRUE)
}
\arguments{
\item{zlim}{height limits over which to compute volume curve}

\item{clip2shp}{TRUE/FALSE decides if raster is to be clippd to a shapefile}

\item{zones}{Number of elevation zones for curves}

\item{change.dir}{Change working directory to directy containing raster}
}
\description{
Plots volume-elevation curve and hypsographic curve
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
