% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{fx.trail}
\alias{fx.trail}
\title{Round and add trailing zeros}
\usage{
fx.trail(x, n = 2)
}
\arguments{
\item{x}{Number to round and add trailing}

\item{n}{Number of decimal points}
}
\value{
x
}
\description{
Round and add trailing zeros
}
\examples{
fx.trail(20,5)
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
