% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{hills.legend}
\alias{hills.legend}
\title{Plot a legend}
\usage{
hills.legend(r, xlim = NULL, ylim = NULL, brks = NA)
}
\arguments{
\item{r}{a raster}

\item{xlim}{xlimits to plot}

\item{ylim}{ylimits to plot}

\item{brks}{optionally provide breaks}
}
\description{
Plot a legend
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}
