% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{hills}
\alias{hills}
\title{Plot terrain with hillsahde}
\usage{
hills(r, xlim = NULL, ylim = NULL, main = "Terreng", legend = TRUE,
  brks = NA)
}
\arguments{
\item{r}{raster}

\item{xlim}{xlimits to plot}

\item{ylim}{ylimits to plot}

\item{main}{Title}

\item{legend}{FALSE/TRUE decide whether to plot a legend. Dont plot a legend if you plan to add more objects to the plot. They wont be plotted properly}

\item{brks}{optionally provide breaks to decide the colorlevels}
}
\description{
Plot terrain with hillsahde
}

