% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{Interactive_terrain_edit}
\alias{Interactive_terrain_edit}
\title{Select Working Directory
Select raster with elevation data
Read help Shapefiles (a point shapefile with POINT_Z attribute)
Read help Shapefiles (Other points,polygons and polylines)
harmonize coordinate system/projection of rasters and help shapefiles
Decide Zoom area to process
Decide number of near parallel editing lines to construct for editing terrin
Decide whether to read elevations of start and endpoints of editing lines from terrain or enter them manually
Call edit Terrain function
Interactive_terrain_edit()}
\usage{
Interactive_terrain_edit()
}
\description{
Select Working Directory
Select raster with elevation data
Read help Shapefiles (a point shapefile with POINT_Z attribute)
Read help Shapefiles (Other points,polygons and polylines)
harmonize coordinate system/projection of rasters and help shapefiles
Decide Zoom area to process
Decide number of near parallel editing lines to construct for editing terrin
Decide whether to read elevations of start and endpoints of editing lines from terrain or enter them manually
Call edit Terrain function
Interactive_terrain_edit()
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

