% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\docType{data}
\name{POINTS_Z}
\alias{POINTS_Z}
\title{Terrain data example}
\format{a point shapefile withe elevations:
\describe{
  \item{POINTS_Z}{Attribute withy heigh (masl)}
}}
\usage{
POINTS_Z
}
\description{
Terrain data example
}
\keyword{datasets}
