% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{polygonFromXY}
\alias{polygonFromXY}
\title{Function to create a SpatialPolygonsDataFrame from a matrix of xy coordinates}
\usage{
polygonFromXY(pts = null, Raster = null, n = 4,
  coord.sys = "+proj=utm +zone=33 +ellps=WGS84 +datum=WGS84 +units=m +no_defs ")
}
\arguments{
\item{pts}{matrix of x,y points}
}
\value{
SpatialPolygonsDataFrame from
}
\description{
Function to create a SpatialPolygonsDataFrame from a matrix of xy coordinates
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

