% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{raster2x11}
\alias{raster2x11}
\title{Function to call x11 with H and Wbased on a raster}
\usage{
raster2x11(r, H = 20, n = 1, xlim = NULL, ylim = NULL)
}
\arguments{
\item{r}{raster}

\item{H}{height}

\item{n}{number of widths}
}
\description{
Function to call x11 with H and Wbased on a raster
}

