% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\docType{data}
\name{River}
\alias{River}
\title{Terrain data example}
\format{A shapefile showing the river banks:
\describe{
  \item{Shapefile}{Polygon or Lines}
}}
\source{
\url{hoydedata.no}
}
\usage{
River
}
\description{
Terrain data example
}
\keyword{datasets}

