% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{sample.line.inkl.start.end.make.dem}
\alias{sample.line.inkl.start.end.make.dem}
\title{Function for drawing points at a given interval along a line
Find the elavtion of points by interpolating between the elevations of the start and endpoints if \code{interpZ==TRUE}
Create a raster by interpolating the elevations}
\usage{
sample.line.inkl.start.end.make.dem(x, sdist = 1, interp.z = TRUE, spdf,
  zmanual = FALSE)
}
\arguments{
\item{x}{A SpatialLinesDataFrame (must have a projection in metres), Lines along which points are created. Must have data with first column (ID) second column (Zs) and third column (Ze) . Zs and Ze are start and end elevations for each line (ID).}

\item{sdist}{interval in metres}

\item{interp.z}{Should be TRUE (for now)}

\item{spdf}{A SpatialPolygonsData Frame (a bounding polygon for the interpolation )}

\item{zmanual}{TRUE/FALSE decide whether to read the elevations from the raster (feks Excavation)or manually enter them (feks filling)}
}
\value{
\code{results} A SpatialPointsDataFrame
}
\description{
Function for drawing points at a given interval along a line
Find the elavtion of points by interpolating between the elevations of the start and endpoints if \code{interpZ==TRUE}
Create a raster by interpolating the elevations
}

