% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{sample.line}
\alias{sample.line}
\title{Function for drawing points at a given interval along a line}
\usage{
sample.line(x, sdist = 100)
}
\arguments{
\item{x}{A SpatialLinesDataFrame (must have a projection in metres)}

\item{sdist}{interval in metres}
}
\value{
\code{results} A SpatialPointsDataFrame
}
\description{
Function for drawing points at a given interval along a line
}

