% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{shadowtext}
\alias{shadowtext}
\title{Function to add text with a halo/shadow}
\usage{
shadowtext(x, y = NULL, labels, col = "white", bg = "black",
  theta = seq(0, 2 * pi, length.out = 50), r = 0.1, ...)
}
\arguments{
\item{x}{x coordinate(s)}

\item{y}{y coordinate(s)}

\item{labels}{labels to add (same lengths as x and y)}

\item{col}{label foreground colour}

\item{bg}{lable halo/shadow color}

\item{theta}{theta controls appearance of shadow}
}
\description{
Function to add text with a halo/shadow
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

