% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\docType{data}
\name{Terrain}
\alias{Terrain}
\title{Terrain data example}
\format{A raster dataset with elevatiopms:
\describe{
  \item{Terrain}{mas}
}}
\source{
\url{hoydedata.no}
}
\usage{
Terrain
}
\description{
Terrain data example
}
\keyword{datasets}

