% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{tk.1.value}
\alias{tk.1.value}
\title{Return a value entered using a tk dialog box}
\usage{
tk.1.value(msg = "Enter number of Parallel Interpolation Lines (minimum 2)",
  varname = "Number of Lines to Draw")
}
\arguments{
\item{msg}{message}

\item{varname}{Number of Lines to draw}
}
\value{
\code{x} a valuye entered
}
\description{
Return a value entered using a tk dialog box
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

