% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{tk.2.values}
\alias{tk.2.values}
\title{Return two numeric values entered using a tk dialog box}
\usage{
tk.2.values(msg = "Enter Start and End Elevations for current Line",
  varname = c("Start", "end"))
}
\arguments{
\item{msg}{message}

\item{varname}{Number of Lines to draw}
}
\value{
\code{c(x,y)} values entered
}
\description{
Return two numeric values entered using a tk dialog box
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

