% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{tk.n.elv}
\alias{tk.n.elv}
\title{Return n number of values entered using a tk dialog box}
\usage{
tk.n.elv(msg = "Enter Line Numbers from left to Right", n = 2)
}
\arguments{
\item{msg}{message}

\item{n}{Number of values to enter and return}
}
\value{
\code{out} values entered
}
\description{
Return n number of values entered using a tk dialog box
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

