% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{tk.n.strings}
\alias{tk.n.strings}
\title{prompts user to Enter strings and returns them to the variable}
\usage{
tk.n.strings(msg = "Enter Values", n = 2, Varnames = NULL)
}
\arguments{
\item{msg}{message}

\item{n}{number of values}

\item{Varnames}{variable names}
}
\value{
out
}
\description{
prompts user to Enter strings and returns them to the variable
}
