% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EditTerrainJEFunctions.R
\name{ui}
\alias{ui}
\title{USER INTERFACE}
\usage{
ui()
}
\arguments{
\item{OutPrefix}{prefix to output filename. Also used as output foldername}
}
\description{
USER INTERFACE
}
\author{
Emmanuel Jjunju, \email{ejjunju@gmail.com}
}

